+++ docarray/array/array/pushpull/file.py
@@ -4,8 +4,11 @@ from typing import Dict, Iterator, List, Optional, Type
 
 from typing_extensions import TYPE_CHECKING
 
-from docarray.array.array.pushpull import ConcurrentPushException, __cache_path__
 from docarray.array.array.pushpull.helpers import _from_binary_stream, _to_binary_stream
+from docarray.array.array.pushpull.pushpull import (
+    ConcurrentPushException,
+    __cache_path__,
+)
 
 if TYPE_CHECKING:
     from docarray import BaseDocument, DocumentArray
